Did a bunch of UI polish and fixed a bug where the Add, Edit and Delete button wouldn't appear for the token's owner.
Started the night with some coinage work, only to realize that the way 5e handles it and the way DCC handles it is probably incompatible.

Also commented out the Totals section and Currency section at the top of the page for now. Honestly, I don't see a lot of value outside total weight... but I'm sure someone else will want it so I might tackle it sooner rather than later.